# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service for investigating \"Deep inspection\" of VMs in VMware vSphere",
        "title": "VM Deep Inspection Demo API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/inspector/v2v": {
            "post": {
                "description": "Queue V2V-only inspection jobs through the selected VDDK or NFC backend. Each job creates and removes a temporary vSphere snapshot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Start asynchronous virt-v2v inspection",
                "parameters": [
                    {
                        "description": "VM names to inspect",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.startV2VRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Jobs accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "A VM already has an active job or the queue is full",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Selected inspection backend is unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Cancel all queued and running V2V inspections",
                "responses": {
                    "202": {
                        "description": "Cancellation requested",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/inspector/v2v/status": {
            "get": {
                "description": "Return the latest persisted V2V job state for each VM.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Get V2V inspection status",
                "responses": {
                    "200": {
                        "description": "V2V status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Status query failed",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/inspector/v2v/{vmName}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Cancel V2V inspection for one VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM name",
                        "name": "vmName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Cancellation requested",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No active job",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.startV2VRequest": {
            "type": "object",
            "required": [
                "vm_names"
            ],
            "properties": {
                "vm_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "VM ID is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
    }
}